Separate CPU & GPU backends into separate files
